Renamed requirements to performances
